refactor: repository 책임 분리하기
