+++ src/main/java/subway/repository/LineRepository.java
@@ -3,7 +3,6 @@ package subway.repository;
 import org.springframework.stereotype.Repository;
 import subway.dao.LineDao;
 import subway.dao.SectionDao;
-import subway.dao.StationDao;
 import subway.domain.line.Color;
 import subway.domain.line.Line;
 import subway.domain.line.Name;
@@ -11,11 +10,8 @@ import subway.domain.section.Section;
 import subway.domain.section.Sections;
 import subway.domain.station.Station;
 import subway.entity.LineEntity;
-import subway.entity.SectionEntity;
 import subway.entity.SectionStationEntity;
-import subway.entity.StationEntity;
 import subway.exception.common.NotFoundLineException;
-import subway.exception.common.NotFoundStationException;
 import subway.exception.line.AlreadyExistLineException;
 
 import java.util.ArrayList;
@@ -26,18 +22,12 @@ import java.util.stream.Collectors;
 @Repository
 public class LineRepository {
 
-    private static final int UPBOUND_STATION_INDEX = 0;
-    private static final int DOWNBOUND_STATION_INDEX = 1;
-    private static final int LAST_SECTION_INDEX = 0;
-
     private final LineDao lineDao;
     private final SectionDao sectionDao;
-    private final StationDao stationDao;
 
-    public LineRepository(final LineDao lineDao, final SectionDao sectionDao, final StationDao stationDao) {
+    public LineRepository(final LineDao lineDao, final SectionDao sectionDao) {
         this.lineDao = lineDao;
         this.sectionDao = sectionDao;
-        this.stationDao = stationDao;
     }
 
     public Line save(final Line line) {
@@ -84,29 +74,6 @@ public class LineRepository {
         return new Line(lineEntity.getId(), lineEntity.getName(), lineEntity.getColor(), sections);
     }
 
-    public List<Station> saveInitStations(final Section section, final Long lineId) {
-        List<StationEntity> stationEntities = List.of(
-                new StationEntity(section.getLeftStation().getName(), lineId),
-                new StationEntity(section.getRightStation().getName(), lineId));
-        List<StationEntity> insertedStationEntities = stationDao.insertInit(stationEntities);
-
-        sectionDao.insert(new SectionEntity(insertedStationEntities.get(UPBOUND_STATION_INDEX).getId(),
-                insertedStationEntities.get(DOWNBOUND_STATION_INDEX).getId(),
-                lineId,
-                section.getDistance()));
-
-        return List.of(
-                new Station(insertedStationEntities.get(UPBOUND_STATION_INDEX).getId(),
-                        insertedStationEntities.get(UPBOUND_STATION_INDEX).getName()),
-                new Station(insertedStationEntities.get(DOWNBOUND_STATION_INDEX).getId(),
-                        insertedStationEntities.get(DOWNBOUND_STATION_INDEX).getName()));
-    }
-
-    public Station saveStation(final Station insertStation, final Long lineId) {
-        StationEntity insertedStation = stationDao.insert(new StationEntity(insertStation.getName(), lineId));
-        return new Station(insertedStation.getId(), insertedStation.getName());
-    }
-
     public List<Line> readAll() {
         List<LineEntity> lineEntities = lineDao.findAll();
         List<SectionStationEntity> sectionStationEntities = sectionDao.findAll();
@@ -133,72 +100,5 @@ public class LineRepository {
             );
         }
         return sections;
-    }
-
-    public Station findByNameAndLineId(String baseStation, Long lineId) {
-        StationEntity stationEntity = stationDao.findByNameAndLineId(baseStation, lineId).orElseThrow(NotFoundStationException::new);
-        return new Station(stationEntity.getId(), stationEntity.getName());
-    }
-
-    public Station findStationById(Long stationId) {
-        StationEntity stationEntity = stationDao.findById(stationId).orElseThrow(NotFoundStationException::new);
-        return new Station(stationEntity.getId(), stationEntity.getName());
-    }
-
-    public Station findStationByName(final String station, final String lineName) {
-        StationEntity stationEntity = stationDao.findByName(station, lineName).orElseThrow(NotFoundStationException::new);
-        return new Station(stationEntity.getId(), stationEntity.getName());
-    }
-
-    public Sections readAllSection() {
-        List<SectionStationEntity> sectionStationEntities = sectionDao.findAll();
-        return new Sections(getSections(sectionStationEntities));
-    }
-
-    public void updateSectionAndDeleteStation(final Long lineId, final Sections updateSections, final Station station) {
-        List<Section> existSections = getSections(sectionDao.findByLineId(lineId));
-        Sections sections = new Sections(existSections);
-        if (updateSections.isEmpty()) {
-            List<Long> stationIds = List.of(sections.findUpBoundStation().getId(), sections.findDownBoundStation().getId());
-            stationDao.deleteBothById(stationIds);
-            sectionDao.deleteById(existSections.get(LAST_SECTION_INDEX).getId());
-            return;
-        }
-        deleteUpdateSection(updateSections, lineId);
-        addUpdateSection(updateSections, lineId);
-        stationDao.deleteById(station.getId());
-    }
-
-    public void updateSection(final Sections sections, final Long lineId) {
-        deleteUpdateSection(sections, lineId);
-        addUpdateSection(sections, lineId);
-    }
-
-    private void deleteUpdateSection(final Sections sections, final Long lineId) {
-        List<Section> updatedSections = sections.getSections();
-        List<Section> existedSections = getSections(sectionDao.findByLineId(lineId));
-        existedSections.removeAll(updatedSections);
-
-        List<Long> sectiondIds = existedSections.stream()
-                .map(Section::getId).
-                collect(Collectors.toList());
-
-        if (sectiondIds.isEmpty()) {
-            return;
-        }
-        sectionDao.deleteBothById(sectiondIds);
-    }
-
-    private void addUpdateSection(final Sections sections, final Long lineId) {
-        List<Section> updatedSections = sections.getSections();
-        List<Section> existedSections = getSections(sectionDao.findByLineId(lineId));
-        updatedSections.removeAll(existedSections);
-
-        sectionDao.insertBoth(updatedSections.stream()
-                .map(section -> new SectionEntity(
-                        section.getLeftStation().getId(),
-                        section.getRightStation().getId(),
-                        lineId, section.getDistance())).
-                collect(Collectors.toList()));
     }
 }
